refactor tidal data fetch script
