+++ scripts/jma/parse.js
@@ -1,0 +1,16 @@
+import { Line } from "../../src/models/DataLine.js";
+
+export function parseTidalData(data) {
+  const lines = data
+    .toString()
+    .split("\n")
+    .filter((line) => line.length > 0)
+    .map((line) => new Line(line, 9));
+
+  const events = lines.reduce((agg, line) => {
+    agg[line.dateString] = line.eventsAsObjects;
+    return agg;
+  }, {});
+
+  return events;
+}
